2025/ANE2/contab-backend/requirements.txt
pathlib==1.0.1
python-crontab==3.3.0
